+++ account_banking_sepa_direct_debit/__init__.py
@@ -1,6 +1,3 @@
 # -*- coding: utf-8 -*-
-# © 2013 Akretion (www.akretion.com)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import models
-from . import wizard
