+++ account_banking_sepa_direct_debit/data/account_payment_method.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+<data noupdate="0">
+
+
+<record id="sepa_direct_debit" model="account.payment.method">
+    <field name="name">SEPA Direct Debit for customers</field>
+    <field name="code">sepa_direct_debit</field>
+    <field name="payment_type">inbound</field>
+    <field name="pain_version">pain.008.001.02</field>
+</record>
+
+
+</data>
+</openerp>
