+++ account_banking_sepa_direct_debit/data/mandate_expire_cron.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <!--
-    Copyright (C) 2013 Akretion (http://www.akretion.com/)
+    Copyright (C) 2013-2016 Akretion (http://www.akretion.com/)
     @author Alexis de Lattre <alexis.delattre@akretion.com>
     The licence is in the file __openerp__.py
 -->
