+++ account_banking_sepa_direct_debit/demo/sepa_direct_debit_demo.xml
@@ -3,13 +3,11 @@
 <openerp>
 <data noupdate="1">
 
-<record id="sepa_direct_debit_mode" model="payment.mode">
-    <field name="name">SEPA Direct Debit La Banque Postale</field>
-    <field name="journal" ref="account.bank_journal"/>
-    <field name="bank_id" ref="account_banking_payment_export.main_company_iban"/>
+<record id="payment_mode_inbound_sepa_dd1" model="account.payment.mode">
+    <field name="name">SEPA Direct Debit of customers</field>
     <field name="company_id" ref="base.main_company"/>
-    <field name="type" ref="export_sdd_008_001_02"/>
-    <field name="purchase_ok" eval="False"/>
+    <field name="bank_account_link">variable</field>
+    <field name="payment_method_id" ref="sepa_direct_debit"/>
     <field name="default_journal_ids" search="[('type', 'in', ('sale', 'sale_refund'))]"/>
 </record>
 
@@ -18,7 +16,7 @@
 </record>
 
 <record id="res_partner_12_mandate" model="account.banking.mandate">
-    <field name="partner_bank_id" ref="account_banking_payment_export.res_partner_12_iban"/>
+    <field name="partner_bank_id" ref="account_payment_mode.res_partner_12_iban"/>
     <field name="format">sepa</field>
     <field name="type">recurrent</field>
     <field name="recurrent_sequence_type">first</field>
@@ -27,7 +25,7 @@
 </record>
 
 <record id="res_partner_2_mandate" model="account.banking.mandate">
-    <field name="partner_bank_id" ref="account_banking_payment_export.res_partner_2_iban"/>
+    <field name="partner_bank_id" ref="account_payment_mode.res_partner_2_iban"/>
     <field name="format">sepa</field>
     <field name="type">recurrent</field>
     <field name="recurrent_sequence_type">first</field>
