+++ account_banking_sepa_direct_debit/models/__init__.py
@@ -3,4 +3,6 @@
 from . import res_company
 from . import account_banking_mandate
 from . import bank_payment_line
-from . import payment_mode
+from . import account_payment_mode
+from . import account_payment_method
+from . import account_payment_order
