+++ account_banking_sepa_direct_debit/models/account_banking_mandate.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2013 Akretion - Alexis de Lattre <alexis.delattre@akretion.com>
+# © 2013-2016 Akretion - Alexis de Lattre <alexis.delattre@akretion.com>
 # © 2014 Serv. Tecnol. Avanzados - Pedro M. Baeza
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
@@ -16,29 +16,13 @@ logger = logging.getLogger(__name__)
 class AccountBankingMandate(models.Model):
     """SEPA Direct Debit Mandate"""
     _inherit = 'account.banking.mandate'
-    _track = {
-        'recurrent_sequence_type': {
-            'account_banking_sepa_direct_debit.recurrent_sequence_type_first':
-            lambda self, cr, uid, obj, ctx=None:
-            obj['recurrent_sequence_type'] == 'first',
-            'account_banking_sepa_direct_debit.'
-            'recurrent_sequence_type_recurring':
-            lambda self, cr, uid, obj, ctx=None:
-            obj['recurrent_sequence_type'] == 'recurring',
-            'account_banking_sepa_direct_debit.recurrent_sequence_type_final':
-            lambda self, cr, uid, obj, ctx=None:
-            obj['recurrent_sequence_type'] == 'final',
-        }
-    }
 
     format = fields.Selection(
-        selection_add=[('sepa', _('Sepa Mandate'))],
-        default='sepa',
-    )
+        selection_add=[('sepa', 'Sepa Mandate')], default='sepa')
     type = fields.Selection([('recurrent', 'Recurrent'),
                              ('oneoff', 'One-Off')],
                             string='Type of Mandate',
-                            track_visibility='always')
+                            track_visibility='onchange')
     recurrent_sequence_type = fields.Selection(
         [('first', 'First'),
          ('recurring', 'Recurring'),
@@ -46,24 +30,10 @@ class AccountBankingMandate(models.Model):
         string='Sequence Type for Next Debit', track_visibility='onchange',
         help="This field is only used for Recurrent mandates, not for "
              "One-Off mandates.", default="first")
-    sepa_migrated = fields.Boolean(
-        string='Migrated to SEPA', track_visibility='onchange',
-        help="If this field is not active, the mandate section of the next "
-             "direct debit file that include this mandate will contain the "
-             "'Original Mandate Identification' and the 'Original Creditor "
-             "Scheme Identification'. This is required in a few countries "
-             "(Belgium for instance), but not in all countries. If this is "
-             "not required in your country, you should keep this field always "
-             "active.", default=True)
-    original_mandate_identification = fields.Char(
-        string='Original Mandate Identification', track_visibility='onchange',
-        size=35,
-        help="When the field 'Migrated to SEPA' is not active, this field "
-             "will be used as the Original Mandate Identification in the "
-             "Direct Debit file.")
-    scheme = fields.Selection([('CORE', 'Basic (CORE)'),
-                               ('B2B', 'Enterprise (B2B)')],
-                              string='Scheme', default="CORE")
+    scheme = fields.Selection([
+        ('CORE', 'Basic (CORE)'),
+        ('B2B', 'Enterprise (B2B)')],
+        string='Scheme', default="CORE", track_visibility='onchange')
     unique_mandate_reference = fields.Char(size=35)  # cf ISO 20022
 
     @api.multi
@@ -74,30 +44,6 @@ class AccountBankingMandate(models.Model):
                     not mandate.recurrent_sequence_type):
                 raise exceptions.Warning(
                     _("The recurrent mandate '%s' must have a sequence type.")
-                    % mandate.unique_mandate_reference)
-
-    @api.multi
-    @api.constrains('type', 'recurrent_sequence_type', 'sepa_migrated')
-    def _check_migrated_to_sepa(self):
-        for mandate in self:
-            if (mandate.type == 'recurrent' and not mandate.sepa_migrated and
-                    mandate.recurrent_sequence_type != 'first'):
-                raise exceptions.Warning(
-                    _("The recurrent mandate '%s' which is not marked as "
-                      "'Migrated to SEPA' must have its recurrent sequence "
-                      "type set to 'First'.")
-                    % mandate.unique_mandate_reference)
-
-    @api.multi
-    @api.constrains('type', 'original_mandate_identification', 'sepa_migrated')
-    def _check_original_mandate_identification(self):
-        for mandate in self:
-            if (mandate.type == 'recurrent' and not mandate.sepa_migrated and
-                    not mandate.original_mandate_identification):
-                raise exceptions.Warning(
-                    _("You must set the 'Original Mandate Identification' on "
-                      "the recurrent mandate '%s' which is not marked as "
-                      "'Migrated to SEPA'.")
                     % mandate.unique_mandate_reference)
 
     @api.multi
@@ -137,5 +83,5 @@ class AccountBankingMandate(models.Model):
                 'The following SDD Mandate IDs has been set to expired: %s'
                 % expired_mandates.ids)
         else:
-            logger.info('0 SDD Mandates must be set to Expired')
+            logger.info('0 SDD Mandates had to be set to Expired')
         return True
