+++ account_banking_sepa_direct_debit/models/bank_payment_line.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2015 Akretion - Alexis de Lattre <alexis.delattre@akretion.com>
+# © 2015-2016 Akretion - Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from openerp import models, api
