+++ account_banking_sepa_direct_debit/models/common.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2013 Akretion (www.akretion.com)
+# © 2013-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # © 2014 Serv. Tecnol. Avanzados - Pedro M. Baeza
 # © 2016 Antiun Ingenieria S.L. - Antonio Espinosa
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
