--- account_banking_sepa_direct_debit/security/original_mandate_required_security.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  Copyright (C) 2013 Akretion (http://www.akretion.com/)
-  @author: Alexis de Lattre <alexis.delattre@akretion.com>
-  The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data noupdate="1">
-
-<record id="group_original_mandate_required" model="res.groups">
-    <field name="name">Original Mandate Required (SEPA)</field>
-    <field name="category_id" ref="base.module_category_hidden"/>
-</record>
-
-</data>
-</openerp>
