+++ account_banking_sepa_direct_debit/views/account_banking_mandate_view.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  Copyright (C) 2013 Akretion (http://www.akretion.com)
+  Copyright (C) 2013-2016 Akretion (http://www.akretion.com)
   @author: Alexis de Lattre <alexis.delattre@akretion.com>
   The licence is in the file __openerp__.py
 -->
@@ -11,7 +11,7 @@
 <data>
 
 
-<record id="sdd_mandate_form" model="ir.ui.view">
+<record id="view_mandate_form" model="ir.ui.view">
     <field name="name">sdd.mandate.form</field>
     <field name="model">account.banking.mandate</field>
     <field name="inherit_id" ref="account_banking_mandate.view_mandate_form"/>
@@ -25,14 +25,10 @@
             <field name="scheme" attrs="{'invisible': [('format', '!=', 'sepa')],
                                          'required': [('format', '=', 'sepa')]}"/>
         </field>
-        <field name="last_debit_date" position="after">
-            <field name="sepa_migrated" groups="account_banking_sepa_direct_debit.group_original_mandate_required"/>
-            <field name="original_mandate_identification" attrs="{'invisible': [('sepa_migrated', '=', True)], 'required': [('sepa_migrated', '=', False)]}" groups="account_banking_sepa_direct_debit.group_original_mandate_required"/>
-        </field>
     </field>
 </record>
 
-<record id="sdd_mandate_tree" model="ir.ui.view">
+<record id="view_mandate_tree" model="ir.ui.view">
     <field name="name">sdd.mandate.tree</field>
     <field name="model">account.banking.mandate</field>
     <field name="inherit_id" ref="account_banking_mandate.view_mandate_tree"/>
@@ -45,7 +41,7 @@
     </field>
 </record>
 
-<record id="sdd_mandate_search" model="ir.ui.view">
+<record id="view_mandate_search" model="ir.ui.view">
     <field name="name">sdd.mandate.search</field>
     <field name="model">account.banking.mandate</field>
     <field name="inherit_id" ref="account_banking_mandate.view_mandate_search"/>
@@ -64,68 +60,6 @@
                 context="{'group_by': 'recurrent_sequence_type'}"/>
         </filter>
     </field>
-</record>
-
-<record id="mandate_action" model="ir.actions.act_window">
-    <field name="name">SEPA Direct Debit Mandates</field>
-    <field name="res_model">account.banking.mandate</field>
-    <field name="view_mode">tree,form</field>
-    <field name="help" type="html">
-        <p class="oe_view_nocontent_create">
-        Click to create a new SEPA Direct Debit Mandate.
-        </p><p>
-        A SEPA Direct Debit Mandate is a document signed by your customer that gives you the autorization to do one or several direct debits on his bank account.
-        </p>
-    </field>
-</record>
-
-<menuitem id="account_banking_mandate.mandate_menu"
-    parent="account_payment.menu_main_payment"
-    action="mandate_action"
-    sequence="20"
-    />
-
-<record id="sdd_mandate_partner_bank_tree" model="ir.ui.view">
-    <field name="name">sdd.mandate.res.partner.bank.tree</field>
-    <field name="model">res.partner.bank</field>
-    <field name="inherit_id" ref="account_banking_mandate.mandate_partner_bank_tree"/>
-    <field name="arch" type="xml">
-        <field name="mandate_ids" position="attributes">
-            <attribute name="string">SDD Mandates</attribute>
-        </field>
-    </field>
-</record>
-
-<record id="sdd_mandate_partner_form" model="ir.ui.view">
-    <field name="name">sdd.mandate.partner.form</field>
-    <field name="model">res.partner</field>
-    <field name="inherit_id" ref="account_banking_mandate.mandate_partner_form"/>
-    <field name="arch" type="xml">
-        <xpath expr="//field[@name='bank_ids']/tree/field[@name='mandate_ids']" position="attributes">
-            <attribute name="string">SDD Mandates</attribute>
-        </xpath>
-    </field>
-</record>
-
-<record id="recurrent_sequence_type_first" model="mail.message.subtype">
-    <field name="name">Sequence Type set to First</field>
-    <field name="res_model">account.banking.mandate</field>
-    <field name="default" eval="False"/>
-    <field name="description">Sequence Type set to First</field>
-</record>
-
-<record id="recurrent_sequence_type_recurring" model="mail.message.subtype">
-    <field name="name">Sequence Type set to Recurring</field>
-    <field name="res_model">account.banking.mandate</field>
-    <field name="default" eval="False"/>
-    <field name="description">Sequence Type set to Recurring</field>
-</record>
-
-<record id="recurrent_sequence_type_final" model="mail.message.subtype">
-    <field name="name">Sequence Type set to Final</field>
-    <field name="res_model">account.banking.mandate</field>
-    <field name="default" eval="False"/>
-    <field name="description">Sequence Type set to Final</field>
 </record>
 
 </data>
