+++ account_banking_sepa_direct_debit/views/account_payment_mode.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- © 2015 Antiun Ingenieria S.L. - Antonio Espinosa
+     © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). -->
+<openerp>
+<data>
+
+<record id="account_payment_mode_form" model="ir.ui.view">
+    <field name="name">Add SEPA identifiers on payment mode form</field>
+    <field name="model">account.payment.mode</field>
+    <field name="inherit_id" ref="account_banking_pain_base.account_payment_mode_form"/>
+    <field name="arch" type="xml">
+        <group name="pain_options" position="inside">
+            <field name="sepa_creditor_identifier" invisible="1"/>
+            <!-- This field should be set visible by localization modules
+            for countries that may have several sepa_creditor_identifier for
+            the same company (I guess only Spain) -->
+        </group>
+    </field>
+</record>
+
+</data>
+</openerp>
