--- account_banking_sepa_direct_debit/views/payment_mode_view.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- © 2015 Antiun Ingenieria S.L. - Antonio Espinosa
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). -->
-<openerp>
-<data>
-
-<record id="view_payment_mode_form_inherit" model="ir.ui.view">
-    <field name="name">Add SEPA identifiers</field>
-    <field name="model">payment.mode</field>
-    <field name="inherit_id" ref="account_banking_pain_base.view_payment_mode_form_inherit"/>
-    <field name="arch" type="xml">
-        <group name="sepa_identifiers" position="inside">
-            <group>
-                <field name="sepa_creditor_identifier"
-                       attrs="{'invisible': [('sepa_type', '!=', 'sepa_direct_debit')]}"/>
-                <field name="original_creditor_identifier"
-                       attrs="{'invisible': [('sepa_type', '!=', 'sepa_direct_debit')]}"/>
-            </group>
-        </group>
-    </field>
-</record>
-
-</data>
-</openerp>
