+++ account_banking_sepa_direct_debit/views/res_company_view.xml
@@ -1,20 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  Copyright (C) 2013 Akretion (http://www.akretion.com)
+  Copyright (C) 2013-2016 Akretion (http://www.akretion.com)
   @author: Alexis de Lattre <alexis.delattre@akretion.com>
   The licence is in the file __openerp__.py
 -->
 <openerp>
 <data>
 
-<record id="sdd_res_company_form" model="ir.ui.view">
+<record id="view_company_form" model="ir.ui.view">
     <field name="name">sepa_direct_debit.res.company.form</field>
     <field name="model">res.company</field>
     <field name="inherit_id" ref="account_banking_pain_base.view_company_form"/>
     <field name="arch" type="xml">
         <group name="pain" position="inside">
             <field name="sepa_creditor_identifier"/>
-            <field name="original_creditor_identifier" groups="account_banking_sepa_direct_debit.group_original_mandate_required"/>
         </group>
     </field>
 </record>
