--- account_banking_sepa_direct_debit/wizard/export_sdd_view.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  Copyright (C) 2010-2012 Akretion (http://www.akretion.com)
-  @author: Alexis de Lattre <alexis.delattre@akretion.com>
-  The licence is in the file __openerp__.py
--->
-<openerp>
-<data>
-
-<record id="banking_export_sdd_wizard_view" model="ir.ui.view">
-    <field name="name">banking.export.sdd.wizard.view</field>
-    <field name="model">banking.export.sdd.wizard</field>
-    <field name="arch" type="xml">
-        <form string="SEPA Direct Debit XML file generation">
-            <field name="state" invisible="True"/>
-            <group states="create">
-                <field name="batch_booking" />
-                <field name="charge_bearer" />
-            </group>
-            <group states="finish">
-                <field name="total_amount" />
-                <field name="nb_transactions" />
-                <field name="file" filename="filename" />
-                <field name="filename" invisible="True"/>
-            </group>
-            <footer>
-                <button string="Generate" name="create_sepa" type="object" class="oe_highlight" states="create"/>
-                <button string="Validate" name="save_sepa" type="object" class="oe_highlight" states="finish"/>
-                <button string="Cancel" special="cancel" class="oe_link"/>
-            </footer>
-        </form>
-    </field>
-</record>
-
-</data>
-</openerp>
